mvp_genai_accounting/backend/requirements.txt
fastapi
uvicorn
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
openai
pydantic
python-dotenv
qdrant-client